Add SimpleHammer as a weapon class for RoleGame prototype
The SimpleHammer.h and SimpleHammer.cpp files add another Weapon subclass
to the RoleGame prototype. The SimpleHammer has the following
attributes:

	- name = "Simple Hammer"
	- hitpoints = 25
	- If the armor is less than 30, the hammer will ignore all armor
	  points

Like the other current weapon classes, the class simply contains the
single "hit" function, to calculate the damage specified by the weapon,
given the amount of armor. In addition are the constructor and
destructor of the class.

No test cases for this class have been added yet to main.cpp.