Add CrazyRandomSword as a weapon class for RoleGame prototype
The CrazyRandomSword.h and CrazyRandomSword.cpp files add another Weapon
subclass to the RoleGame prototype. The CrazyRandomSword has the
following attributes:

	- name = "Crazy Random Sword"
	- hitpoints = random number from 7-100
	- ignores random integer amount of armor points from 2-one third of
	  the armor

Like the other weapon classes, this class simply contains the "hit"
function, to calculate the damage specified by the weapon, given the
amount of armor and randomly generated number. Included is also the
constructor and desctructor.

In addition to these added files, the WeaponFactory.cpp file has been
modified to change its "getWeapon" function to include cases for the
previous class SimpleHammer, as well as the newly added class
"CrazyRandomSword". This is where the random number from 7-100 is
calculated.