Add LuckyGauntlet as a weapon class for RoleGame prototype
The LuckyGauntlet.h and LuckyGauntlet.cpp files add another Weapon
subclass to the RoleGame prototype. The LuckyGauntlet has the following
attributes:

	- name = "Lucky Gauntlet"
	- hitpoints = 50
	- when the function 'hit' is called, one of three options will
	  be selected to change the damage the weapon does. Two of the
	  options are not so good, while the third is very rewarding,
	  hence the name of the weapon 'lucky'

The first option does not ignore any armor and lowers the damage of the
gauntlet to 1. The second option ignores all armor, but still keeps the
damage of the weapon to 1. The third, and best, option doubles the
initial hitpoints of the weapon (50), and then picks a random number
between 0 and the armor level to ignore.

Lastly, the WeaponFactory.cpp file has been modified to include this
weapon as 'gauntlet'.